requirements.txt
kazoo
PyYAML
scutils